Revert "Merge pull request #374 from alphagov/PP-1948/asyn-capture-feature-flag"
This reverts commit 3ff461011abd5c0f8280cf66b5de0aea96d5603e, reversing
changes made to 56053b451e74c1fdea589f89550959cf2cf04bf0.
