+++ src/main/java/uk/gov/pay/connector/app/ConnectorConfiguration.java
@@ -19,10 +19,6 @@ public class ConnectorConfiguration extends Configuration {
     @NotNull
     private ExecutorServiceConfig executorServiceConfig = new ExecutorServiceConfig();
 
-    @NotNull
-    @JsonProperty
-    private boolean asynchronousCapture;
-
     @Valid
     @NotNull
     private CaptureProcessConfig captureProcessConfig;
@@ -135,9 +131,5 @@ public class ConnectorConfiguration extends Configuration {
 
     public CaptureProcessConfig getCaptureProcessConfig() {
         return captureProcessConfig;
-    }
-
-    public boolean isAsynchronousCapture() {
-        return asynchronousCapture;
     }
 }
