+++ src/main/java/uk/gov/pay/connector/service/CardCaptureService.java
@@ -23,7 +23,6 @@ import static uk.gov.pay.connector.model.domain.ChargeStatus.*;
 public class CardCaptureService extends CardService implements TransactionalGatewayOperation<BaseCaptureResponse> {
 
     private static List<ChargeStatus> legalStatuses = ImmutableList.of(
-            AUTHORISATION_SUCCESS,
             CAPTURE_APPROVED,
             CAPTURE_APPROVED_RETRY
     );
